Moved template to model
