+++ src/main/se/kth/iv1350/model/TotalRevenue.java
@@ -1,4 +1,4 @@
-package se.kth.iv1350.view;
+package se.kth.iv1350.model;
 
 import se.kth.iv1350.model.Amount;
 import se.kth.iv1350.model.CashRegisterObserver;
